set aliases more careful.

+++ .bash_aliases
@@ -31,34 +31,40 @@ alias npw="pwgen -Bs 16 1"
 alias godeps='go get ./..'
 
 # Docker aliases
-alias drm="docker rm $(docker ps -a -q)"
-alias drmi="docker rmi $(docker images -q)"
-alias dclean="drm && drmi"
+if hash docker 2>/dev/null; then
+  alias drm="docker rm $(docker ps -a -q)"
+  alias drmi="docker rmi $(docker images -q)"
+  alias dclean="drm && drmi"
+fi
 
 # Git aliases
-alias gl="git log --pretty=oneline --abbrev-commit --graph --decorate"
-alias gd="git diff"
-alias ga="git add"
-alias gst="git status"
-alias gs="git status -sb"
-alias gbd="git branch -D"
-alias gca="git commit -a -m"
-alias gm="git merge --no-ff"
-alias gpt="git push --tags"
-alias gp="git push"
-alias gpa="git push --all"
-alias gu="git checkout -- ."
-alias grh="git reset --hard"
-alias gco="git checkout"
-alias gba="git branch -a"
-alias gfp="git fetch origin --prune"
-alias gr='cd "$(git rev-parse --show-toplevel)"'
+if hash git 2>/dev/null; then
+  alias gl="git log --pretty=oneline --abbrev-commit --graph --decorate"
+  alias gd="git diff"
+  alias ga="git add"
+  alias gst="git status"
+  alias gs="git status -sb"
+  alias gbd="git branch -D"
+  alias gca="git commit -a -m"
+  alias gm="git merge --no-ff"
+  alias gpt="git push --tags"
+  alias gp="git push"
+  alias gpa="git push --all"
+  alias gu="git checkout -- ."
+  alias grh="git reset --hard"
+  alias gco="git checkout"
+  alias gba="git branch -a"
+  alias gfp="git fetch origin --prune"
+  alias gr='cd "$(git rev-parse --show-toplevel)"'
+fi
 
 # Vagrant aliases
-alias vup="vagrant up"
-alias vha="vagrant halt"
-alias vst="vagrant status"
-alias vdf="vagrant destroy -f; rm -rf .vagrant"
+if hash vagrant 2>/dev/null; then
+  alias vup="vagrant up"
+  alias vha="vagrant halt"
+  alias vst="vagrant status"
+  alias vdf="vagrant destroy -f; rm -rf .vagrant"
+fi
 
 # cd into new directory
 function take {
